requirements.txt
larcv
hydra-core
pytorch-lightning
torch-optimizer
torch-geometric
sparseconvnet